+++ frontend/src/utils/index.ts
@@ -1,5 +1,5 @@
 export { parseNumberAndUnit, divideSize, multiplySize } from './calculateUtils';
 export { calculatePosition } from './arrangement';
-export { getRemainingSeconds } from './time';
+export { getRemainingSeconds, convertSecToHHMMSS } from './time';
 export { convertArrayToObject } from './participantUtils';
 export { checkYoutubeURLValidation, getYoutubeEmbedURL, isShorts } from './youtube';
